refactor: trim media understanding runner test imports
