+++ src/media-understanding/runner.skip-tiny-audio.test.ts
@@ -5,12 +5,8 @@ import { describe, expect, it } from "vitest";
 import type { MsgContext } from "../auto-reply/templating.js";
 import type { OpenClawConfig } from "../config/config.js";
 import { MIN_AUDIO_FILE_BYTES } from "./defaults.js";
-import {
-  buildProviderRegistry,
-  createMediaAttachmentCache,
-  normalizeMediaAttachments,
-  runCapability,
-} from "./runner.js";
+import { createMediaAttachmentCache, normalizeMediaAttachments } from "./runner.attachments.js";
+import { buildProviderRegistry, runCapability } from "./runner.js";
 import type { AudioTranscriptionRequest } from "./types.js";
 
 async function withAudioFixture(params: {
