+++ src/media-understanding/runner.test-utils.ts
@@ -3,7 +3,7 @@ import os from "node:os";
 import path from "node:path";
 import { withEnvAsync } from "../test-utils/env.js";
 import { MIN_AUDIO_FILE_BYTES } from "./defaults.js";
-import { createMediaAttachmentCache, normalizeMediaAttachments } from "./runner.js";
+import { createMediaAttachmentCache, normalizeMediaAttachments } from "./runner.attachments.js";
 
 type MediaFixtureParams = {
   ctx: { MediaPath: string; MediaType: string };
